pyproject.toml
[tool.poetry]
name = "nonebot-plugin-dailysign"
version = "0.0.2"
description = "一个简单的每日签到插件"
authors = ["kexue <x@kexue.io>"]
license = "MIT"
homepage = "https://github.com/kexue-z/nonebot-plugin-dailysign"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-tortoise-orm = "^0.0.1-alpha.2"


[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}
flake8 = "^4.0.1"
isort = "^5.10.1"
nb-cli = "^0.6.6"
click = "^8.1.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
